+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add Volume's download link to `patterns/template-taxonomu-volumes.php`
+
 ## [0.4.0] 2025-07-13
 ### Added
 - Add `question` style for interview-like post